requirements.txt
requests==2.22.0
slackclient==2.5.0
pyowm==2.10.0
googletrans==2.4.0
oxforddictionaries==0.1.101
numpy==1.17.3
google_images_download==2.6.1
bs4==0.0.1
aiohttp==3.6.2
apscheduler==3.6.3
nest-asyncio==1.3.3
